fix(sandbox): bound policy-hint probes (#5978)
Split detection, rendering, and emission into focused modules. Cap runtime probes at one second and stop retrying failed reads so best-effort guidance cannot delay or replace the command result.
